+++ common/prboom/z_bmalloc.c
@@ -54,7 +54,7 @@ inline static void* getelem(bmalpool_t *p, size_t size, size_t n)
 inline static PUREFUNC int iselem(const bmalpool_t *pool, size_t size, const void* p)
 {
   // CPhipps - need portable # of bytes between pointers
-  int dif = (const char*)p - (const char*)pool;
+  int dif = (int)((const char*)p - (const char*)pool);
 
   dif -= sizeof(bmalpool_t);
   dif -= pool->blocks;
@@ -71,7 +71,7 @@ void* Z_BMalloc(struct block_memory_alloc_s *pzone)
   while (*pool != NULL) {
     byte *p = memchr((*pool)->used, unused_block, (*pool)->blocks); // Scan for unused marker
     if (p) {
-      int n = p - (*pool)->used;
+      long n = p - (*pool)->used;
 #ifdef SIMPLECHECKS
       if ((n<0) || ((size_t)n>=(*pool)->blocks))
   I_Error("Z_BMalloc: memchr returned pointer outside of array");
